ipam/multi-pool: Move manager logic into its own file
In order to improve the reusability of the multi-pool package, let's
clearly separate the parts specific for pod IPAM, like the Allocator and
the local node allocations CIDRs synchronization, and the ones that
implement the core logic, like the manager.

This allows to reuse the multi-pool IPAM core logic in the context of
the upcoming Cilium Network driver.
